Include out-of-window measurements in process_skew result
In the process_skew test, we are only interested in the start times
relative to the start time of the first process, so time windows have no
meaning.
We want to keep track of how many measurements would have been valid,
however, as indications on how to set wait times and window sizes.